realestate_monitoring/requirements.txt
scrapy>=2.8.0
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3.0
APScheduler>=3.9
streamlit>=1.30
plotly>=5.15
python-dotenv>=1.0
SQLAlchemy>=2.0
scrapy-playwright>=0.1.0  # optional, for dynamic content
